apps/guru-web/guru-api/requirements.txt
# GURU API Backend Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

